feat(auth): improve token refresh flow and middleware
- Fixed token refresh loop in middleware
- Moved token refresh logic to authRequests.ts
- Improved error handling for expired tokens
- Updated cookie settings for better security
- Simplified middleware response handling
